security jwt, comments controllers... and events service

+++ backend17/src/main/java/com/ocio/backend17/controllers/CommentsController.java
@@ -6,10 +6,14 @@ import com.ocio.backend17.dto.ResponseMessage;
 import com.ocio.backend17.dto.ResponseMessageWithIndex;
 import com.ocio.backend17.entities.Comments;
 
+import com.ocio.backend17.security.ExtractHeaderData;
+import com.ocio.backend17.security.JWTUtil;
 import com.ocio.backend17.services.ICommentsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,33 +23,41 @@ import java.util.List;
 public class CommentsController {
     @Autowired
     ICommentsImpl iCommentsImpl;
+    @Autowired
+    ExtractHeaderData extractHeaderData;
 
+    @PreAuthorize("hasAuthority('create:comments')")
     @PostMapping(value="/api/comments", consumes = "application/json")
     @ResponseBody
-    ResponseEntity<?> createOrUpdateComment(@RequestBody String jsonComment) throws JsonProcessingException {
+    ResponseEntity<?> createOrUpdateComment(@RequestBody String jsonComment, @RequestHeader HttpHeaders headers) throws JsonProcessingException {
         ObjectMapper om = new ObjectMapper();
         Comments comment=om.readValue(jsonComment, Comments.class);
         if(!(comment.getEvent_id()>0)){
             return new ResponseEntity<>(new ResponseMessage("Fields cannot be empty"), HttpStatus.BAD_REQUEST);
         }else{
+            comment.setAssistant(extractHeaderData.extractJWTUsername(headers));
             return new ResponseEntity<>(iCommentsImpl.addComment(comment), HttpStatus.CREATED);
         }
     }
+
+    @PreAuthorize("hasAuthority('read:comments')")
     @GetMapping("/api/comments/byevent/{event_id}")
     ResponseEntity<List<Comments>> getAll(@PathVariable("event_id") Double event_id){
         return new ResponseEntity<>(iCommentsImpl.findByEventId(event_id), HttpStatus.OK);
 
     }
 
+    @PreAuthorize("hasAuthority('delete:comments')")
     @DeleteMapping("/api/user/{comment_id}/{index}")
     @ResponseBody
-    ResponseEntity<?> deleteByEmail(@PathVariable("comment_id") Double id, @PathVariable("index") int index){
-        if(iCommentsImpl.findbyId(id).isPresent()) {
+    ResponseEntity<?> deleteByEmail(@PathVariable("comment_id") Double id, @PathVariable("index") int index,  @RequestHeader HttpHeaders headers){
+        if(iCommentsImpl.findbyId(id).isPresent() && iCommentsImpl.findbyId(id).get().getAssistant().equals(extractHeaderData.extractJWTUsername(headers))) {
             iCommentsImpl.deleteById(id);
             return new ResponseEntity<>(new ResponseMessageWithIndex("User deleted", index), HttpStatus.NO_CONTENT);
-        }else{
-            return new ResponseEntity<>(new ResponseMessage("User not found"), HttpStatus.NO_CONTENT);
+        }else if(iCommentsImpl.findbyId(id).isPresent() && !(iCommentsImpl.findbyId(id).get().getAssistant().equals(extractHeaderData.extractJWTUsername(headers)))){
+            return new ResponseEntity<>(new ResponseMessage("Only author cans delete his own comments"), HttpStatus.UNAUTHORIZED);
         }
+        return new ResponseEntity<>(new ResponseMessage("Comment id not found"), HttpStatus.NO_CONTENT);
     }
 
 }
